refactor: 패키지 분리 - aws/s3 분리 #32
